MonitoringDataStorage: permanent folder in ctor

+++ projects/copper-monitoring/copper-monitoring-core/src/main/java/org/copperengine/monitoring/core/data/MonitoringDataStorage.java
@@ -216,7 +216,9 @@ public class MonitoringDataStorage {
 
     private static File createFolder(String dirPath) throws IOException {        
         File dir = new File(dirPath);
-        dir.delete();
+        if(dir.isFile()) {
+            dir.delete();
+        }
         dir.mkdirs();
         if(!dir.exists() || !dir.isDirectory()) {
             throw new IOException("Cannot create directory " + dirPath);
